Extract gRPC stubs to separate project
