+++ .gitignore
@@ -1,3 +1,4 @@
 **/target/
 **/.*
+**/*.iml
 !.gitignore
