--- service-consumer-spring/src/main/resources/grpc/blogposts.proto
@@ -1,53 +1,0 @@
-syntax = "proto3";
-import "google/protobuf/empty.proto";
-option java_package = "de.samples.apicomparison.consumer.clients.grpc.stub";
-option java_multiple_files = true;
-
-/*
- *  S I M P L E   T Y P E S
- */
-message UUID {
-  string value = 1;
-}
-
-message Timestamp {
-  int64 seconds = 1;
-  int32 nanos = 2;
-}
-
-message Author {
-  UUID id = 1;
-  string name = 2;
-  optional string description = 3;
-}
-
-message BlogPost {
-  UUID id = 1;
-  string title = 2;
-  string content = 3;
-  Timestamp timestamp = 4;
-  optional Author author = 5;
-  repeated string hashTags = 6;
-}
-
-/*
- *  S E R V I C E
- */
-message ListOfBlogPosts {
-  repeated BlogPost blogPost = 1;
-}
-
-message BlogPostInput {
-  string title = 2;
-  string content = 3;
-  optional UUID author_id = 5;
-  repeated string hashTags = 6;
-}
-
-service BlogPostService {
-  rpc FindAllBlogPosts(google.protobuf.Empty) returns (ListOfBlogPosts);
-  rpc FindAllBlogPostsStream(google.protobuf.Empty) returns (stream BlogPost);
-  rpc CreateBlogPost(BlogPostInput) returns (BlogPost);
-  rpc CreateBlogPostStream(stream BlogPostInput) returns (stream BlogPost);
-  rpc DeleteBlogPost(UUID) returns (google.protobuf.Empty);
-}
