+++ service-provider/src/main/java/de/samples/apicomparison/provider/boundary/grpc/mappers/UuidMapper.java
@@ -7,16 +7,16 @@ import java.util.UUID;
 @Mapper(componentModel = "spring")
 public interface UuidMapper {
 
-  default de.samples.apicomparison.provider.boundary.grpc.stub.UUID map(UUID source) {
+  default de.samples.apicomparison.stubs.grpc.UUID map(UUID source) {
     if (null == source) {
       return null;
     }
-    return de.samples.apicomparison.provider.boundary.grpc.stub.UUID.newBuilder()
+    return de.samples.apicomparison.stubs.grpc.UUID.newBuilder()
       .setValue(source.toString())
       .build();
   }
 
-  default UUID map(de.samples.apicomparison.provider.boundary.grpc.stub.UUID source) {
+  default UUID map(de.samples.apicomparison.stubs.grpc.UUID source) {
     if (null == source || source.getValue().isEmpty()) {
       return null;
     }
